+++ me/dawars/CraftingPillars/blocks/CraftingPillarBlock.java
@@ -3,7 +3,6 @@ package me.dawars.CraftingPillars.blocks;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
 import me.dawars.CraftingPillars.CraftingPillars;
-import me.dawars.CraftingPillars.tiles.TileEntityBrewingPillar;
 import me.dawars.CraftingPillars.tiles.TileEntityCraftingPillar;
 import net.minecraft.block.material.Material;
 import net.minecraft.client.renderer.texture.IconRegister;
@@ -13,7 +12,6 @@ import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.ChatMessageComponent;
 import net.minecraft.util.Icon;
 import net.minecraft.util.MathHelper;
 import net.minecraft.world.World;
@@ -24,32 +22,32 @@ public class CraftingPillarBlock extends BaseBlockContainer
 	{
 		super(id, mat);
 	}
-	
+
 	@Override
 	public int getRenderType()
 	{
 		return CraftingPillars.craftingPillarRenderID;
 	}
-	
+
 	@Override
 	public boolean renderAsNormalBlock()
 	{
 		return false;
 	}
-	
+
 	@Override
 	public boolean isOpaqueCube()
 	{
 		return false;
 	}
-	
+
 	@Override
 	public void onBlockClicked(World world, int x, int y, int z, EntityPlayer player)
 	{
 		if(!world.isRemote)
 		{
 			TileEntityCraftingPillar workTile = (TileEntityCraftingPillar) world.getBlockTileEntity(x, y, z);
-			
+
 			if(player.isSneaking())
 			{
 				while(workTile.getStackInSlot(workTile.getSizeInventory()) != null)
@@ -61,28 +59,28 @@ public class CraftingPillarBlock extends BaseBlockContainer
 			}
 		}
 	}
-	
+
 	@Override
 	public void onBlockPlacedBy(World world, int x, int y, int z, EntityLivingBase entity, ItemStack stack)
 	{
 		world.setBlockMetadataWithNotify(x, y, z, determineOrientation(world, x, y, z, entity), 0);
 	}
-	
+
 	public static int determineOrientation(World world, int x, int y, int z, EntityLivingBase entity)
 	{
-		return MathHelper.floor_double((double) (entity.rotationYaw * 4.0F / 360.0F) + 0.5D) & 3;
+		return MathHelper.floor_double(entity.rotationYaw * 4.0F / 360.0F + 0.5D) & 3;
 	}
-	
+
 	@Override
 	public boolean onBlockActivated(World world, int x, int y, int z, EntityPlayer player, int side, float hitX, float hitY, float hitZ)
 	{
 		if(world.isRemote)
 			return true;
-		
+
 		TileEntityCraftingPillar workTile = (TileEntityCraftingPillar) world.getBlockTileEntity(x, y, z);
 
 		ItemStack equipped = player.getCurrentEquippedItem();
-		
+
 		//Thaumcraft start
 		if(CraftingPillars.modThaumcraft)
 		{
@@ -95,7 +93,7 @@ public class CraftingPillarBlock extends BaseBlockContainer
 						ItemStack in = equipped.copy();
 						if(!player.capabilities.isCreativeMode)
 							equipped.stackSize--;
-						
+
 						in.stackSize = 1;
 						workTile.setInventorySlotContents(10, in);
 						return true;
@@ -108,13 +106,13 @@ public class CraftingPillarBlock extends BaseBlockContainer
 			}
 		}
 		//Thaumcraft end
-		
+
 		if(hitY < 1F && !player.isSneaking())
 		{
 			workTile.showNum = !workTile.showNum;
 			workTile.onInventoryChanged();
 		}
-		
+
 		if(hitY == 1F)
 		{
 			if(player.isSneaking())
@@ -129,9 +127,9 @@ public class CraftingPillarBlock extends BaseBlockContainer
 					hitZ = 2;
 				if(hitZ < 0)
 					hitZ = 0;
-				
+
 				int i = (int) (hitX * 3 + hitZ);
-				
+
 				if(workTile.getStackInSlot(i) != null)
 					workTile.dropItemFromSlot(i, player);
 			}
@@ -147,16 +145,16 @@ public class CraftingPillarBlock extends BaseBlockContainer
 					hitZ = 2;
 				if(hitZ < 0)
 					hitZ = 0;
-				
+
 				int i = (int) (hitX * 3 + hitZ);
-				
+
 				if(workTile.getStackInSlot(i) == null)
 				{
-					
+
 					ItemStack in = equipped.copy();
 					if(!player.capabilities.isCreativeMode)
 						equipped.stackSize--;
-					
+
 					in.stackSize = 1;
 					workTile.setInventorySlotContents(i, in);
 				}
@@ -164,71 +162,71 @@ public class CraftingPillarBlock extends BaseBlockContainer
 				{
 					if(!player.capabilities.isCreativeMode)
 						equipped.stackSize--;
-					
+
 					workTile.decrStackSize(i, -1);
 				}
 			}
 		}
 		return true;
 	}
-	
+
 	@Override
 	public void breakBlock(World world, int x, int y, int z, int par5, int par6)
 	{
 		if(!world.isRemote)
 		{
-//			TileEntityCraftingPillar workTile = (TileEntityCraftingPillar) world.getBlockTileEntity(x, y, z);
-//			for(int i = 0; i < 3; i++)
-//			{
-//				for(int k = 0; k < 3; k++)
-//				{
-//					if(workTile.getStackInSlot(i * 3 + k) != null)
-//					{
-//						EntityItem itemDropped = new EntityItem(world, x + 0.1875D + i * 0.3125D, y + 1D, z + 0.1875D + k * 0.3125D, workTile.getStackInSlot(i * 3 + k));
-//						itemDropped.motionX = itemDropped.motionY = itemDropped.motionZ = 0D;
-//						
-//						if(workTile.getStackInSlot(i * 3 + k).hasTagCompound())
-//							itemDropped.getEntityItem().setTagCompound((NBTTagCompound) workTile.getStackInSlot(i * 3 + k).getTagCompound().copy());
-//						
-//						world.spawnEntityInWorld(itemDropped);
-//					}
-//				}
-//			}
-			
+			//			TileEntityCraftingPillar workTile = (TileEntityCraftingPillar) world.getBlockTileEntity(x, y, z);
+			//			for(int i = 0; i < 3; i++)
+			//			{
+			//				for(int k = 0; k < 3; k++)
+			//				{
+			//					if(workTile.getStackInSlot(i * 3 + k) != null)
+			//					{
+			//						EntityItem itemDropped = new EntityItem(world, x + 0.1875D + i * 0.3125D, y + 1D, z + 0.1875D + k * 0.3125D, workTile.getStackInSlot(i * 3 + k));
+			//						itemDropped.motionX = itemDropped.motionY = itemDropped.motionZ = 0D;
+			//						
+			//						if(workTile.getStackInSlot(i * 3 + k).hasTagCompound())
+			//							itemDropped.getEntityItem().setTagCompound((NBTTagCompound) workTile.getStackInSlot(i * 3 + k).getTagCompound().copy());
+			//						
+			//						world.spawnEntityInWorld(itemDropped);
+			//					}
+			//				}
+			//			}
+
 			TileEntityCraftingPillar pillarTile = (TileEntityCraftingPillar) world.getBlockTileEntity(x, y, z);
-			
+
 			for(int i = 0; i < pillarTile.getSizeInventory() + 2; i++)
 			{
 				if(pillarTile.getStackInSlot(i) != null)
 				{
 					EntityItem itemDropped = new EntityItem(world, x + 0.5D, y, z + 0.5D, pillarTile.getStackInSlot(i));
 					itemDropped.motionX = itemDropped.motionY = itemDropped.motionZ = 0D;
-					
+
 					if(pillarTile.getStackInSlot(i).hasTagCompound())
 						itemDropped.getEntityItem().setTagCompound((NBTTagCompound) pillarTile.getStackInSlot(i).getTagCompound().copy());
-					
+
 					world.spawnEntityInWorld(itemDropped);
 				}
 			}
 		}
-		
+
 		super.breakBlock(world, x, y, z, par5, par6);
 	}
-	
+
 	@Override
 	public TileEntity createNewTileEntity(World world)
 	{
 		TileEntityCraftingPillar tile = new TileEntityCraftingPillar();
 		return tile;
 	}
-	
+
 	@Override
 	@SideOnly(Side.CLIENT)
 	public void registerIcons(IconRegister itemIcon)
 	{
 		this.blockIcon = itemIcon.registerIcon(CraftingPillars.id + ":craftingPillar_side");
 	}
-	
+
 	@Override
 	@SideOnly(Side.CLIENT)
 	public Icon getIcon(int par1, int par2)
