+++ me/dawars/CraftingPillars/blocks/PotPillarBlock.java
@@ -5,30 +5,16 @@ import java.util.Random;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
 import me.dawars.CraftingPillars.CraftingPillars;
-import me.dawars.CraftingPillars.tiles.TileEntityCraftingPillar;
 import me.dawars.CraftingPillars.tiles.TileEntityPotPillar;
-import me.dawars.CraftingPillars.world.gen.ChristmasTreeGen;
-import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
-import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.texture.IconRegister;
-import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.item.EntityItem;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.ChatMessageComponent;
 import net.minecraft.util.Icon;
-import net.minecraft.util.MathHelper;
 import net.minecraft.world.World;
-import net.minecraft.world.gen.feature.WorldGenBigTree;
-import net.minecraft.world.gen.feature.WorldGenForest;
-import net.minecraft.world.gen.feature.WorldGenHugeTrees;
-import net.minecraft.world.gen.feature.WorldGenTaiga2;
-import net.minecraft.world.gen.feature.WorldGenTrees;
-import net.minecraft.world.gen.feature.WorldGenerator;
-import net.minecraftforge.event.terraingen.TerrainGen;
 
 public class PotPillarBlock extends BaseBlockContainer
 {
@@ -37,25 +23,25 @@ public class PotPillarBlock extends BaseBlockContainer
 		super(id, mat);
 		this.setTickRandomly(true);
 	}
-	
+
 	@Override
 	public int getRenderType()
 	{
 		return CraftingPillars.potPillarRenderID;
 	}
-	
+
 	@Override
 	public boolean renderAsNormalBlock()
 	{
 		return false;
 	}
-	
+
 	@Override
 	public boolean isOpaqueCube()
 	{
 		return false;
 	}
-	
+
 	/**
 	 * Used during tree growth to determine if newly generated leaves can replace this block.
 	 *
@@ -70,7 +56,7 @@ public class PotPillarBlock extends BaseBlockContainer
 	{
 		return false;
 	}
-	
+
 	@Override
 	@SideOnly(Side.CLIENT)
 	/**
@@ -80,13 +66,13 @@ public class PotPillarBlock extends BaseBlockContainer
 	{
 		return true;
 	}
-	
+
 	@Override
 	public boolean onBlockActivated(World world, int x, int y, int z, EntityPlayer player, int side, float hitX, float hitY, float hitZ)
 	{
 		if(world.isRemote)
 			return true;
-		
+
 		TileEntityPotPillar pillarTile = (TileEntityPotPillar) world.getBlockTileEntity(x, y, z);
 
 		if(hitY < 1F && !player.isSneaking() /*&& player.getCurrentEquippedItem() == null*/)
@@ -94,7 +80,7 @@ public class PotPillarBlock extends BaseBlockContainer
 			pillarTile.showNum = !pillarTile.showNum;
 			pillarTile.onInventoryChanged();
 		}
-		
+
 		if(hitY == 1F)
 		{
 			if(player.isSneaking())//pick out
@@ -109,7 +95,7 @@ public class PotPillarBlock extends BaseBlockContainer
 					{
 						if(!player.capabilities.isCreativeMode)
 							player.getCurrentEquippedItem().stackSize--;
-						
+
 						ItemStack in = player.getCurrentEquippedItem().copy();
 						in.stackSize = 1;
 						pillarTile.setInventorySlotContents(0, in);
@@ -119,36 +105,36 @@ public class PotPillarBlock extends BaseBlockContainer
 		}
 		return true;
 	}
-	
+
 	@Override
 	public void breakBlock(World world, int x, int y, int z, int par5, int par6)
 	{
 		if(!world.isRemote)
 		{
 			TileEntityPotPillar workTile = (TileEntityPotPillar) world.getBlockTileEntity(x, y, z);
-			
+
 			if(workTile.getStackInSlot(0) != null)
 			{
 				EntityItem itemDropped = new EntityItem(world, x + 0.1875D, y + 1D, z + 0.1875D, workTile.getStackInSlot(0));
 				itemDropped.motionX = itemDropped.motionY = itemDropped.motionZ = 0D;
-				
+
 				if(workTile.getStackInSlot(0).hasTagCompound())
 					itemDropped.getEntityItem().setTagCompound((NBTTagCompound) workTile.getStackInSlot(0).getTagCompound().copy());
-				
+
 				world.spawnEntityInWorld(itemDropped);
 			}
 		}
-		
+
 		super.breakBlock(world, x, y, z, par5, par6);
 	}
-	
+
 	@Override
 	public TileEntity createNewTileEntity(World world)
 	{
 		TileEntityPotPillar tile = new TileEntityPotPillar();
 		return tile;
 	}
-	
+
 	@Override
 	public void updateTick(World world, int x, int y, int z, Random rand)
 	{
@@ -176,7 +162,7 @@ public class PotPillarBlock extends BaseBlockContainer
 					double d2 = rand.nextGaussian() * 0.02D;
 					float width = 0.3F;
 					float height = 0.5F;
-					world.spawnParticle("happyVillager", x + 0.6F + (double)(rand.nextFloat() * width  * 2.0F) - (double)width, y + 1.2D + (double)(rand.nextFloat() * height), z + 0.6F + (double)(rand.nextFloat() * width * 2.0F) - (double)width, d0, d1, d2);
+					world.spawnParticle("happyVillager", x + 0.6F + (double)(rand.nextFloat() * width  * 2.0F) - width, y + 1.2D + rand.nextFloat() * height, z + 0.6F + (double)(rand.nextFloat() * width * 2.0F) - width, d0, d1, d2);
 				}
 			}
 		}
@@ -188,7 +174,7 @@ public class PotPillarBlock extends BaseBlockContainer
 	{
 		this.blockIcon = itemIcon.registerIcon(CraftingPillars.id + ":craftingPillar_side");
 	}
-	
+
 	@Override
 	@SideOnly(Side.CLIENT)
 	public Icon getIcon(int par1, int par2)
