+++ me/dawars/CraftingPillars/network/packets/PacketInGameClick.java
@@ -5,20 +5,18 @@ import java.io.DataOutputStream;
 import java.io.IOException;
 
 import me.dawars.CraftingPillars.CraftingPillars;
-import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.network.packet.Packet250CustomPayload;
-import net.minecraft.server.MinecraftServer;
 
 public class PacketInGameClick extends PillarPacket
 {
 	public int mouseButton, btnId, x, y, z;
-	
+
 	public PacketInGameClick(Packet250CustomPayload packet)
 	{
 		super(CraftingPillars.channelGame);
 		this.receive(packet);
 	}
-	
+
 	public PacketInGameClick(int mouseButton, int button, int x, int y, int z)
 	{
 		super(CraftingPillars.channelGame);
@@ -42,8 +40,8 @@ public class PacketInGameClick extends PillarPacket
 	@Override
 	public void readPacketData(DataInputStream in) throws IOException
 	{
-		this.mouseButton = (int)in.readByte();
-		this.btnId = (int)in.readByte();
+		this.mouseButton = in.readByte();
+		this.btnId = in.readByte();
 		this.x = in.readInt();
 		this.y = in.readInt();
 		this.z = in.readInt();
