+++ me/dawars/CraftingPillars/network/packets/PacketInGuiClick.java
@@ -5,20 +5,18 @@ import java.io.DataOutputStream;
 import java.io.IOException;
 
 import me.dawars.CraftingPillars.CraftingPillars;
-import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.network.packet.Packet250CustomPayload;
-import net.minecraft.server.MinecraftServer;
 
 public class PacketInGuiClick extends PillarPacket
 {
 	public int slot;
-	
+
 	public PacketInGuiClick(Packet250CustomPayload packet)
 	{
 		super(CraftingPillars.channelGui);
 		this.receive(packet);
 	}
-	
+
 	public PacketInGuiClick(int slot)
 	{
 		super(CraftingPillars.channelGui);
@@ -34,7 +32,7 @@ public class PacketInGuiClick extends PillarPacket
 	@Override
 	public void readPacketData(DataInputStream in) throws IOException
 	{
-		this.slot = (int)in.readByte();
+		this.slot = in.readByte();
 	}
 
 	@Override
